add log persistence

+++ nidus/log.py
@@ -1,7 +1,16 @@
 import io
 import json
+import os
 from struct import pack, unpack
 from typing import NamedTuple
+
+
+class PageFullException(Exception):
+    pass
+
+
+class PageEmptyException(Exception):
+    pass
 
 
 class Pager:
@@ -25,51 +34,45 @@ class Pager:
     def __init__(self, log_file):
         self.page_size = 2048
         self.log_file = log_file
-        self.cached_last_page = None
 
-        with open(log_file, "wb+"):
-            pass
+        if not os.path.exists(log_file):
+            with open(log_file, "wb+"):
+                pass
 
-    def get_last_page(self):
-        if self.cached_last_page:
-            return self.cached_last_page
-
-        page_num = 0
-        remaining_space = 0
-        data = None
-
-        for n, space, page in self:
-            page_num = n
-            remaining_space = space
-            data = page
-
-        if data is None:
-            remaining_space = self.page_size - 2
-            data = b"\x00" * remaining_space
-
-        self.cached_last_page = (page_num, remaining_space, data)
-        return page_num, remaining_space, data
-
-    def append_log_entry(self, entry):
-        serialized = entry.to_bytes()
-        serialized_size = len(serialized)
-        page_num, remaining_space, data = self.get_last_page()
-
-        data = data[: self.page_size - 2 - remaining_space] + serialized
-        remaining_space -= serialized_size
-        data += b"\x00" * remaining_space
-
+    def read(self, page_num):
         with open(self.log_file, "rb+") as f:
             f.seek(self.page_size * page_num)
-            f.write(pack(">H", remaining_space) + data)
+            raw_page = f.read(self.page_size)
 
-        self.cached_last_page = (page_num, remaining_space, data)
+        if raw_page:
+            remaining_space = unpack(">H", raw_page[:2])[0]
+            # remove buffer space with null bytes
+            data = page[2 : self.page_size - remaining_space]
+            page = Page(page_num, remaining_space, data)
+        else:
+            page = None
+
+        return page
+
+    def write(self, page):
+        with open(self.log_file, "rb+") as f:
+            f.seek(self.page_size * page.num)
+            raw_page = pack(">H", page.remaining_space) + page.data
+            # buffer remaining space with null bytes
+            raw_page += b"\x00" * page.remaining_space
+            f.write(raw_page)
+
+    def truncate(self, page):
+        # deletes the page and all following pages!
+
+        with open(self.log_file, "rb+") as f:
+            f.seek(self.page_size * page.num)
+            f.truncate()
 
     def __iter__(self):
         page_num = 0
         with open(self.log_file, "rb") as f:
             while True:
-                f.seek(self.page_size * page_num)
                 page = f.read(self.page_size)
 
                 if not page:
@@ -77,8 +80,123 @@ class Pager:
 
                 remaining_space = unpack(">H", page[:2])[0]
                 data = page[2:]
-                yield page_num, remaining_space, data
+                yield Page(page_num, remaining_space, data)
                 page_num += 1
+
+
+class Page:
+    def __init__(self, num, remaining_space, data):
+        self.num = num
+        self.remaining_space = remaining_space
+        self.data = data
+        self.log = []
+        self.load()
+
+    def load(self):
+        cursor = 0
+        end = len(self.data) - self.remaining_space
+        while cursor < end:
+            term, item_size = unpack(">LL", self.data[cursor : cursor + 8])
+            cursor += 8
+            item = self.data[cursor : cursor + item_size]
+            cursor += item_size
+
+            entry = LogEntry(term, json.loads(item))
+            self.log.append(entry)
+
+    def append(self, entry):
+        serialized = entry.to_bytes()
+
+        if len(serialized) > self.remaining_space:
+            raise PageFullException()
+
+        self.data += serialized
+        self.remaining_space -= len(serialized)
+        self.log.append(entry)
+
+    def pop(self):
+        if len(self.log) == 0:
+            raise PageEmptyException()
+
+        entry = self.log.pop()
+        serialized = entry.to_bytes()
+        self.data = self.data[: -len(serialized)]
+        self.remaining_space += len(serialized)
+
+
+class Log:
+    def __init__(self, log_file):
+        self._pager = Pager(log_file)
+        self.pages = []
+        self.load()
+
+    def load(self):
+        for page in self._pager:
+            self.pages.append(page)
+
+        if len(self.pages) == 0:
+            self.pages = [Page(0, self._pager.page_size - 2, b"")]
+
+    def append(self, entry):
+        page = self.pages[-1]
+        try:
+            page.append(entry)
+        except PageFullException:
+            page = Page(page.num + 1, self._pager.page_size - 2, b"")
+            page.append(entry)
+            self.pages.append(page)
+
+        self._pager.write(page)
+
+    def pop(self):
+        page = self.pages[-1]
+        try:
+            page.pop()
+        except PageEmptyException:
+            page = self.pages.pop()
+            self._pager.truncate(page)
+            page = self.pages[-1]
+            page.pop()
+
+        self._pager.write(page)
+
+    def as_list(self):
+        log = []
+        for page in self.pages:
+            log += page.log
+        return log
+
+    def __getitem__(self, idx):
+        if isinstance(idx, slice):
+            # only supports [n:] style slices
+            items = []
+            start = idx.start
+            for page in self.pages:
+                if len(page.log) > start:
+                    items.append(page.log[start])
+                else:
+                    start -= len(page.log)
+            return items
+
+        real_idx = idx
+        for page in self.pages:
+            if len(page.log) > real_idx:
+                return page.log[real_idx]
+            else:
+                real_idx -= len(page.log)
+        raise IndexError("index out of range")
+
+    def __len__(self):
+        size = 0
+        for page in self.pages:
+            size += len(page.log)
+        return size
+
+    def __repr__(self):
+        log = []
+        for page in self.pages:
+            log += page.log
+        return str(log)
 
 
 class LogEntry:
@@ -95,10 +213,6 @@ class LogEntry:
     def to_bytes(self):
         item_json = json.dumps(self.item, separators=(",", ":")).encode("utf8")
         return pack(">LL", self.term, len(item_json)) + item_json
-
-    @classmethod
-    def from_bytes(self):
-        pass
 
 
 def append_entries(log, prev_index, prev_term, entries):
@@ -125,18 +239,8 @@ def apply_all_entries(log, prev_index, entries):
     if len(log) > prev_index + 1:
         clear_upto(log, prev_index + 1)
 
-    for i, entry in enumerate(entries, start=1):
-        # The entry already exists so we are overriding as the leader may
-        # be sending the same entry because it never recieved our response
-        # (the idempotency property).
-        if len(log) > prev_index + i:
-            if log[prev_index + i].term < entry.term:
-                log.append(entry)
-            else:
-                log[prev_index + i] = entry
-        # New entry, append as normal
-        else:
-            log.append(entry)
+    for entry in entries:
+        log.append(entry)
 
 
 def clear_upto(log, upto):
